Moved repository into separate folder
