+++ src/model/index.ts
@@ -1,17 +1,13 @@
 import {
   BaseParams,
-  Game,
-  GameDefinition,
-  GameId,
   GramPerPsgrKm,
   MioPsgrKm,
   MioTons,
   Percent,
   TWh,
   WritableBaseParams,
-} from "./types"
-import { startYear } from "./constants"
-import { API } from "./api"
+} from "../types"
+import { startYear } from "../constants"
 
 export const defaultValues: WritableBaseParams = {
   co2budget: 6700, // https://www.umweltrat.de/SharedDocs/Downloads/DE/01_Umweltgutachten/2016_2020/2020_Umweltgutachten_Kap_02_Pariser_Klimaziele.pdf?__blob=publicationFile&v=22
@@ -55,7 +51,7 @@ export const defaultValues: WritableBaseParams = {
   co2emissionsOthers: 9.243,
 }
 
-const initialGame = {
+export const initialGame = {
   currentYear: startYear,
   values: defaultValues,
   acceptedLaws: [],
@@ -138,87 +134,6 @@ export function createBaseValues(values: WritableBaseParams): BaseParams {
         this.co2emissionsAgriculture +
         this.co2emissionsOthers
       )
-    },
-  }
-}
-
-const unsavedGameId = "00000"
-
-export function initGame(game: GameDefinition = initialGame, id = unsavedGameId): Game {
-  return {
-    id,
-    currentYear: game.currentYear,
-    acceptedLaws: game.acceptedLaws,
-    proposedLaws: game.proposedLaws,
-    rejectedLaws: game.rejectedLaws,
-    values: createBaseValues(game.values),
-    events: [],
-    over: false,
-  }
-}
-
-interface Logger {
-  warn: (msg: string, details?: unknown) => void
-}
-
-interface Storage {
-  setItem: (name: string, value: string) => void
-  getItem: (name: string) => string | null
-}
-
-export default function({
-  api,
-  logger = console,
-  storage = localStorage,
-}: {
-  api: API
-  logger?: Logger
-  storage?: Storage
-}) {
-  return {
-    async createGame(game: GameDefinition = initialGame): Promise<Game> {
-      const newGame = initGame(game)
-      try {
-        return await api.createGame(newGame)
-      } catch (error) {
-        logger.warn("Cannot save new game - trying again later", error)
-        newGame.id = unsavedGameId
-        return newGame
-      }
-    },
-
-    async loadGame(id: GameId): Promise<Game> {
-      const storedItem = storage.getItem("game")
-      if (storedItem) {
-        const storedGame = JSON.parse(storedItem)
-        if (storedGame.id === unsavedGameId || storedGame.id === id) {
-          return initGame(storedGame, id)
-        }
-      }
-
-      try {
-        const storedGame = await api.loadGame(id)
-        return initGame(storedGame, id)
-      } catch (error) {
-        logger.warn(
-          "No game found in localStorage, but the id cannot be found on server either... so no chance to load it."
-        )
-      }
-      throw Error("Game not found")
-    },
-
-    async saveGame(game: Game): Promise<Game> {
-      storage.setItem("game", JSON.stringify(game))
-      try {
-        const savedGame = await api.saveGame({ ...game, id: game.id === unsavedGameId ? "" : game.id })
-        game.id = savedGame.id
-        storage.setItem("game", JSON.stringify(game)) // store again to keep id
-      } catch (error) {
-        logger.warn(
-          "save on server failed - at least the game is saved in localStorage, so you can save it maybe next time"
-        )
-      }
-      return game
     },
   }
 }
